+++ PullThroughDoc/PullThroughDoc.Test/GeneralTests.cs
@@ -1,28 +1,29 @@
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using PullThroughDoc.CodeFixes;
 
-namespace PullThroughDoc.Test
+namespace PullThroughDoc.Test;
+
+[TestClass]
+public class GeneralTests : PullThroughDocCodeFixVerifier
 {
-	[TestClass]
-	public class GeneralTests : PullThroughDocCodeFixVerifier
+	protected override CodeFixProvider CodeFixProvider => new PullThroughDocCodeFixProvider();
+
+	//No diagnostics expected to show up
+	[TestMethod]
+	public void Doesnt_Trigger_For_Nothing()
 	{
-		protected override CodeFixProvider CodeFixProvider => new PullThroughDocCodeFixProvider();
+		var test = @"";
 
-		//No diagnostics expected to show up
-		[TestMethod]
-		public void Doesnt_Trigger_For_Nothing()
-		{
-			var test = @"";
+		VerifyCSharpDiagnostic(test);
+	}
 
-			VerifyCSharpDiagnostic(test);
-		}
-
-		[TestMethod]
-		public void Regions_Documentation_ExcludingRegion()
-		{
-			var test = @"
-    namespace ConsoleApplication1
-    {
+	[TestMethod]
+	public void Regions_Documentation_ExcludingRegion()
+	{
+		var test = @"
+	namespace ConsoleApplication1
+	{
 		class BaseClass 
 		{
 			#region Properties
@@ -30,16 +31,16 @@ namespace PullThroughDoc.Test
 			public virtual string DoThing() { }
 			#endregion
 		}
-        class TypeName : BaseClass
-        {   
+		class TypeName : BaseClass
+		{   
 			public override string DoThing() {}
-        }
-    }";
-			ExpectPullThroughDiagnosticAt(test, "DoThing", 13, 27);
+		}
+	}";
+		ExpectPullThroughDiagnosticAt(test, "DoThing", 13, 27);
 
-			var fixtest = @"
-    namespace ConsoleApplication1
-    {
+		var fixtest = @"
+	namespace ConsoleApplication1
+	{
 		class BaseClass 
 		{
 			#region Properties
@@ -47,21 +48,21 @@ namespace PullThroughDoc.Test
 			public virtual string DoThing() { }
 			#endregion
 		}
-        class TypeName : BaseClass
-        {   
+		class TypeName : BaseClass
+		{   
 			/// <summary>Does A Thing </summary>
 			public override string DoThing() {}
-        }
-    }";
-			VerifyCSharpFix(test, fixtest);
 		}
+	}";
+		VerifyCSharpFix(test, fixtest);
+	}
 
-		[TestMethod]
-		public void Parameters_DocumentationIncluded()
-		{
-			var test = @"
-    namespace ConsoleApplication1
-    {
+	[TestMethod]
+	public void Parameters_DocumentationIncluded()
+	{
+		var test = @"
+	namespace ConsoleApplication1
+	{
 		class BaseClass 
 		{
 			/// <summary>
@@ -72,16 +73,16 @@ namespace PullThroughDoc.Test
 			/// <returns></returns>
 			public virtual string DoThing(string variable) { }
 		}
-        class TypeName : BaseClass
-        {   
+		class TypeName : BaseClass
+		{   
 			public override string DoThing(string variable) {}
-        }
-    }";
+		}
+	}";
 
 
-			var fixtest = @"
-    namespace ConsoleApplication1
-    {
+		var fixtest = @"
+	namespace ConsoleApplication1
+	{
 		class BaseClass 
 		{
 			/// <summary>
@@ -92,8 +93,8 @@ namespace PullThroughDoc.Test
 			/// <returns></returns>
 			public virtual string DoThing(string variable) { }
 		}
-        class TypeName : BaseClass
-        {   
+		class TypeName : BaseClass
+		{   
 			/// <summary>
 			/// Returns a thing
 			/// </summary>
@@ -101,9 +102,8 @@ namespace PullThroughDoc.Test
 			/// <param name=""targetMember""></param>
 			/// <returns></returns>
 			public override string DoThing(string variable) {}
-        }
-    }";
-			VerifyCSharpFix(test, fixtest);
 		}
+	}";
+		VerifyCSharpFix(test, fixtest);
 	}
 }
