+++ PullThroughDoc/PullThroughDoc.Test/InterfaceDocumentationTests.cs
@@ -1,22 +1,23 @@
 ﻿using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using PullThroughDoc.CodeFixes;
 
-namespace PullThroughDoc.Test
+namespace PullThroughDoc.Test;
+
+/// <summary>
+/// Verifies a bunch of tests specific to interfaces
+/// </summary>
+[TestClass]
+public class InterfaceDocumentationTests : PullThroughDocCodeFixVerifier
 {
-	/// <summary>
-	/// Verifies a bunch of tests specific to interfaces
-	/// </summary>
-	[TestClass]
-	public class InterfaceDocumentationTests : PullThroughDocCodeFixVerifier
+
+	protected override CodeFixProvider CodeFixProvider => new PullThroughDocCodeFixProvider();
+
+	//Diagnostic and CodeFix both triggered and checked for
+	[TestMethod]
+	public void Interface_Documentation_PullsThrough()
 	{
-
-		protected override CodeFixProvider CodeFixProvider => new PullThroughDocCodeFixProvider();
-
-		//Diagnostic and CodeFix both triggered and checked for
-		[TestMethod]
-		public void Interface_Documentation_PullsThrough()
-		{
-			var test = @"
+		var test = @"
     namespace ConsoleApplication1
     {
 		interface IInterface 
@@ -30,9 +31,9 @@ namespace PullThroughDoc.Test
         }
     }";
 
-			ExpectPullThroughDiagnosticAt(test, "DoThing", 11, 18);
+		ExpectPullThroughDiagnosticAt(test, "DoThing", 11, 18);
 
-			var fixtest = @"
+		var fixtest = @"
     namespace ConsoleApplication1
     {
 		interface IInterface 
@@ -46,69 +47,68 @@ namespace PullThroughDoc.Test
 			public string DoThing() {}
         }
     }";
-			VerifyCSharpFix(test, fixtest);
-		}
-
-		[TestMethod]
-		public void Interface_MultiLineDocumentation_PullsThrough()
-		{
-			var test = @"
-    namespace ConsoleApplication1
-    {
-		interface IInterface 
-		{
-			/// <summary>Does A Thing </summary>
-			/// <param name=""param1"">parameter</param>
-			/// <returns>A string</returns>
-			string DoThing(string param1);
-		}
-        class TypeName : IInterface
-        {   
-			public string DoThing(string param1) {}
-        }
-    }";
-
-			ExpectPullThroughDiagnosticAt(test, "DoThing", 13, 18);
-
-			var fixtest = @"
-    namespace ConsoleApplication1
-    {
-		interface IInterface 
-		{
-			/// <summary>Does A Thing </summary>
-			/// <param name=""param1"">parameter</param>
-			/// <returns>A string</returns>
-			string DoThing(string param1);
-		}
-        class TypeName : IInterface
-        {   
-			/// <summary>Does A Thing </summary>
-			/// <param name=""param1"">parameter</param>
-			/// <returns>A string</returns>
-			public string DoThing(string param1) {}
-        }
-    }";
-			VerifyCSharpFix(test, fixtest);
-		}
-
-		[TestMethod]
-		public void Interface_WithoutDocumentation_NoAnalyzer()
-		{
-			var test = @"
-    namespace ConsoleApplication1
-    {
-		interface IInterface 
-		{
-			string DoThing();
-		}
-        class TypeName : IInterface
-        {   
-			public string DoThing() {}
-        }
-    }";
-
-			VerifyCSharpDiagnostic(test);
-		}
-
+		VerifyCSharpFix(test, fixtest);
 	}
+
+	[TestMethod]
+	public void Interface_MultiLineDocumentation_PullsThrough()
+	{
+		var test = @"
+    namespace ConsoleApplication1
+    {
+		interface IInterface 
+		{
+			/// <summary>Does A Thing </summary>
+			/// <param name=""param1"">parameter</param>
+			/// <returns>A string</returns>
+			string DoThing(string param1);
+		}
+        class TypeName : IInterface
+        {   
+			public string DoThing(string param1) {}
+        }
+    }";
+
+		ExpectPullThroughDiagnosticAt(test, "DoThing", 13, 18);
+
+		var fixtest = @"
+    namespace ConsoleApplication1
+    {
+		interface IInterface 
+		{
+			/// <summary>Does A Thing </summary>
+			/// <param name=""param1"">parameter</param>
+			/// <returns>A string</returns>
+			string DoThing(string param1);
+		}
+        class TypeName : IInterface
+        {   
+			/// <summary>Does A Thing </summary>
+			/// <param name=""param1"">parameter</param>
+			/// <returns>A string</returns>
+			public string DoThing(string param1) {}
+        }
+    }";
+		VerifyCSharpFix(test, fixtest);
+	}
+
+	[TestMethod]
+	public void Interface_WithoutDocumentation_NoAnalyzer()
+	{
+		var test = @"
+    namespace ConsoleApplication1
+    {
+		interface IInterface 
+		{
+			string DoThing();
+		}
+        class TypeName : IInterface
+        {   
+			public string DoThing() {}
+        }
+    }";
+
+		VerifyCSharpDiagnostic(test);
+	}
+
 }
